chore: renamed protocol files
