+++ src/components/features/protocol/credentials/view/components/credential-title-and-image.tsx
