+++ src/components/features/protocol/credentials/view/components/image-view.tsx
