+++ src/components/features/protocol/data-models/view/components/table-schema.tsx
