+++ src/components/features/protocol/index.ts
@@ -1,5 +1,5 @@
 import CredentialProtocolCreate from './credentials/create/credential-create';
-import CredentialProtocolShow from './credentials/show/credential-show';
+import CredentialProtocolView from './credentials/view/credential-view';
 import Protocol from './protocol';
 
-export { Protocol, CredentialProtocolShow, CredentialProtocolCreate };
+export { Protocol, CredentialProtocolView, CredentialProtocolCreate };
