+++ src/pages/model/[id]/credentials.tsx
@@ -2,14 +2,14 @@ import useTranslation from 'next-translate/useTranslation';
 import dynamic from 'next/dynamic';
 import { useRouter } from 'next/router';
 
-import { DataModelLayout } from '@/components/features/protocol/data-models/show/layout';
+import { DataModelLayout } from '@/components/features/protocol/data-models/view/layout';
 import { IColumnGrid } from '@/components/organisms/data-grid/data-grid';
 import { query } from '@/constants/queries';
 
 const GridViewTab = dynamic(
   () =>
     import(
-      '@/components/features/protocol/data-models/show/components/grid-view-tab'
+      '@/components/features/protocol/data-models/view/components/grid-view-tab'
     ),
   { ssr: false }
 );
