+++ src/pages/model/[id]/index.tsx
@@ -1,13 +1,13 @@
 import { InferGetServerSidePropsType, GetServerSidePropsContext } from 'next';
 import dynamic from 'next/dynamic';
 
-import { DataModelLayout } from '@/components/features/protocol/data-models/show/layout';
+import { DataModelLayout } from '@/components/features/protocol/data-models/view/layout';
 import { gatewayProtocolSDK } from '@/services/gateway-protocol/api';
 import { gqlAnonMethods } from '@/services/hasura/api';
 const OverviewTab = dynamic(
   () =>
     import(
-      '@/components/features/protocol/data-models/show/components/overview-tab'
+      '@/components/features/protocol/data-models/view/components/overview-tab'
     ),
   { ssr: false }
 );
