+++ src/pages/model/[id]/recipients.tsx
@@ -3,13 +3,13 @@ import { useRouter } from 'next/router';
 
 import { IColumnGrid } from '@/components/organisms/data-grid/data-grid';
 import { DataModelsTab } from '@/components/features/explore';
-import { DataModelLayout } from '@/components/features/protocol/data-models/show/layout';
+import { DataModelLayout } from '@/components/features/protocol/data-models/view/layout';
 import { query } from '@/constants/queries';
 
 const GridViewTab = dynamic(
   () =>
     import(
-      '@/components/features/protocol/data-models/show/components/grid-view-tab'
+      '@/components/features/protocol/data-models/view/components/grid-view-tab'
     ),
   { ssr: false }
 );
